+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
  - Rewrite rules are now applied before the request URL is checked for dots.
+ - Rewrite rules can be defined as functions to have greater control over the `dot rule`.
 
 ## v1.0.0
 This version introduces a fair amount of breaking changes. Specifically, instances of the historyApiFallback need to be created via the exported function. Previously, this was not necessary.